Move AuditTrail attributes to seprate subdirectory
