+++ src/DoctrineListener/DoctrineAuditListener.php
@@ -11,9 +11,9 @@ use Doctrine\ORM\Event\PreRemoveEventArgs;
 use Doctrine\ORM\Events;
 use Doctrine\ORM\Mapping\Id;
 use ReflectionClass;
-use SumoCoders\FrameworkCoreBundle\Attribute\AuditTrailDisplayData;
-use SumoCoders\FrameworkCoreBundle\Attribute\AuditTrailIdentifier;
-use SumoCoders\FrameworkCoreBundle\Attribute\AuditTrailSensitiveData;
+use SumoCoders\FrameworkCoreBundle\Attribute\AuditTrail\AuditTrailDisplayData;
+use SumoCoders\FrameworkCoreBundle\Attribute\AuditTrail\AuditTrailIdentifier;
+use SumoCoders\FrameworkCoreBundle\Attribute\AuditTrail\AuditTrailSensitiveData;
 use SumoCoders\FrameworkCoreBundle\Enum\EventAction;
 use SumoCoders\FrameworkCoreBundle\Logger\AuditLogger;
 use Symfony\Component\Serializer\SerializerInterface;
